@@ -12,6 +12,7 @@
   },
   "private": true,
   "dependencies": {
+    "@abacritt/angularx-social-login": "^2.5.1",
     "@abp/ng.account": "~9.3.2",
     "@abp/ng.components": "~9.3.2",
     "@abp/ng.core": "~9.3.2",
@@ -22,32 +23,35 @@
     "@abp/ng.tenant-management": "~9.3.2",
     "@abp/ng.theme.lepton-x": "~4.3.2",
     "@abp/ng.theme.shared": "~9.3.2",
-    "@angular/animations": "~20.0.0",
-    "@angular/common": "~20.0.0",
-    "@angular/compiler": "~20.0.0",
-    "@angular/core": "~20.0.0",
-    "@angular/forms": "~20.0.0",
-    "@angular/localize": "~20.0.0",
-    "@angular/platform-browser": "~20.0.0",
-    "@angular/platform-browser-dynamic": "~20.0.0",
-    "@angular/router": "~20.0.0",
+    "@angular/animations": "~20.3.9",
+    "@angular/common": "~20.3.9",
+    "@angular/compiler": "~20.3.9",
+    "@angular/core": "~20.3.9",
+    "@angular/forms": "~20.3.9",
+    "@angular/localize": "~20.3.9",
+    "@angular/platform-browser": "~20.3.9",
+    "@angular/platform-browser-dynamic": "~20.3.9",
+    "@angular/router": "~20.3.9",
     "@babel/runtime": "^7.28.4",
+    "@ngx-validate/core": "^0.2.0",
     "@swimlane/ngx-datatable": "^22.0.0",
+    "html2canvas": "^1.4.1",
+    "jspdf": "^3.0.3",
     "rxjs": "~7.8.0",
     "tslib": "^2.0.0",
     "zone.js": "~0.15.0"
   },
   "devDependencies": {
     "@abp/ng.schematics": "~9.3.2",
-    "@angular-devkit/build-angular": "~20.0.0",
+    "@angular-devkit/build-angular": "~20.3.8",
     "@angular-eslint/builder": "~20.0.0",
     "@angular-eslint/eslint-plugin": "~20.0.0",
     "@angular-eslint/eslint-plugin-template": "~20.0.0",
     "@angular-eslint/schematics": "~20.0.0",
     "@angular-eslint/template-parser": "~20.0.0",
-    "@angular/cli": "~20.0.0",
-    "@angular/compiler-cli": "~20.0.0",
-    "@angular/language-service": "~20.0.0",
+    "@angular/cli": "~20.3.8",
+    "@angular/compiler-cli": "~20.3.9",
+    "@angular/language-service": "~20.3.9",
     "@types/jasmine": "~3.6.0",
     "@types/node": "^20.0.0",
     "@typescript-eslint/eslint-plugin": "^7.0.0",
