@@ -53,6 +53,36 @@ export class TranslationService {
       'profile.upgrade_vip_account': 'Nâng cấp tài khoản VIP',
       'profile.activate_gift': 'Kích hoạt quà tặng',
       
+      // Notifications
+      'notifications.title': 'Thông báo',
+      'notifications.mark_all_read': 'Đánh dấu là đã đọc',
+      'notifications.no_notifications': 'Không có thông báo mới',
+      
+      // Sidebar
+      'sidebar.verification': 'Tài khoản xác thực',
+      'sidebar.not_verified': 'Chưa xác thực',
+      'sidebar.dashboard': 'Bảng tin',
+      'sidebar.insights': 'TopCV Insights',
+      'sidebar.rewards': 'TopCV Rewards',
+      'sidebar.redeem_gifts': 'Đổi quà',
+      'sidebar.toppy_ai': 'Toppy AI - Đề xuất',
+      'sidebar.suggested_cv': 'CV đề xuất',
+      'sidebar.recruitment_campaigns': 'Chiến dịch tuyển dụng',
+      'sidebar.job_postings': 'Tin tuyển dụng',
+      'sidebar.cv_management': 'Quản lý CV',
+      'sidebar.recruitment_reports': 'Báo cáo tuyển dụng',
+      'sidebar.buy_services': 'Mua dịch vụ',
+      'sidebar.account_settings': 'Cài đặt tài khoản',
+      'sidebar.activity_history': 'Lịch sử hoạt động',
+      
+      // Verification
+      'verify.step_phone': 'Xác thực số điện thoại',
+      'verify.step_company': 'Cập nhật thông tin công ty',
+      'verify.step_license': 'Cập nhật Giấy đăng ký doanh nghiệp',
+      'verify.step_first_job': 'Đăng tin tuyển dụng đầu tiên',
+      'verify.next_step_phone': 'Xác thực số điện thoại',
+      'verify.complete': 'Hoàn thành',
+      
       // Job Page
       'job_page.title': 'Việc làm',
       'job_page.category_placeholder': 'Danh mục Nghề',
@@ -255,137 +285,6 @@ export class TranslationService {
       'pagination.pages': 'trang',
       'profile.logout': 'Đăng xuất',
       
-      // Notifications
-      'notifications.title': 'Thông báo',
-      'notifications.mark_all_read': 'Đánh dấu là đã đọc',
-      'notifications.sample_text': 'Ứng viên cùng ngành với bạn cũng quan tâm vị trí này:',
-
-      // CV Management
-      'cv_management.title': 'Quản lý CV',
-      'cv_management.subtitle': 'Tạo và quản lý CV của bạn một cách dễ dàng',
-      'cv_management.cv_list_title': 'CV đã tạo trên VCareer',
-      'cv_management.create_cv': 'Tạo CV',
-      'cv_management.create_first_cv': 'Tạo CV đầu tiên',
-      'cv_management.no_cv_title': 'Chưa có CV nào',
-      'cv_management.no_cv_description': 'Hãy tạo CV đầu tiên để bắt đầu tìm kiếm việc làm',
-      'cv_management.updated': 'Cập nhật',
-      'cv_management.view': 'Xem',
-      'cv_management.edit': 'Chỉnh sửa',
-      'cv_management.duplicate': 'Sao chép',
-      'cv_management.set_default': 'Đặt mặc định',
-      'cv_management.delete': 'Xóa',
-      'cv_management.confirm_delete': 'Bạn có chắc chắn muốn xóa CV này?',
-      'cv_management.welcome_back': 'Chào bạn trở lại',
-      'cv_management.verified_account': 'Tài khoản đã xác thực',
-      'cv_management.upgrade_account': 'Nâng cấp tài khoản',
-      'cv_management.job_search_settings': 'Cài đặt tìm việc',
-      'cv_management.job_search_status': 'Đang Tắt tìm việc',
-      'cv_management.job_search_description': 'Bật tìm việc giúp hồ sơ của bạn nổi bật hơn và được chú ý nhiều hơn trong danh sách tìm kiếm của NTD.',
-      'cv_management.allow_recruiter_search': 'Cho phép NTD tìm kiếm hồ sơ',
-      'cv_management.contact_methods_title': 'Khi có cơ hội việc làm phù hợp, NTD sẽ liên hệ và trao đổi với bạn qua:',
-      'cv_management.top_connect': 'Nhắn tin qua Top Connect trên VCareer',
-      'cv_management.email_phone': 'Email và Số điện thoại của bạn',
-      'cv_management.download_app': 'Tải App VCareer ngay!',
-      'cv_management.app_description': 'Để không bỏ lỡ bất cứ cơ hội nào từ Nhà tuyển dụng',
-      'cv_management.career_guidance': 'Định hướng nghề nghiệp theo tử vi',
-      'cv_management.career_feature_1': 'Hiểu rõ tổng quan sự nghiệp',
-      'cv_management.career_feature_2': 'Định hướng nghề nghiệp theo tử vi',
-      'cv_management.career_feature_3': 'Làm chủ vận mệnh cùng chuyên gia',
-      'cv_management.register_now': 'Đăng ký ngay',
-      'cv_management.hide_profile': 'Ẩn hồ sơ với NTD',
-      'cv_management.new': 'Mới',
-      'cv_management.hide_profile_description': 'Tôi không muốn CV của tôi hiển thị với danh sách các NTD có tên miền email và thuộc các công ty dưới đây:',
-      'cv_management.default': 'Mặc định',
-      'cv_management.loading': 'Đang tải...',
-      'cv_management.creating_cv': 'Đang tạo CV...',
-      'cv_management.updated_successfully': 'CV đã được cập nhật thành công!',
-      'cv_management.deleted_successfully': 'CV đã được xóa thành công!',
-      'cv_management.delete_failed': 'Xóa CV thất bại',
-      'cv_management.duplicated_successfully': 'CV đã được sao chép thành công!',
-      'cv_management.duplicate_failed': 'Sao chép CV thất bại',
-    'cv_management.set_default_successfully': 'Đặt CV làm mặc định thành công!',
-    'cv_management.set_default_failed': 'Đặt CV làm mặc định thất bại',
-    'cv_management.download': 'Tải về',
-    'cv_management.push_to_top': 'Đẩy top',
-
-    // Download CV Modal
-    'download_cv.title': 'Tải CV',
-    'download_cv.without_logo_title': 'Tải CV không kèm biểu tượng @VCareer',
-    'download_cv.without_logo_description': 'Không giới hạn lượt tải, số CV và số mẫu thiết kế trong vòng 24 giờ.',
-    'download_cv.without_logo_button': 'Tải CV không có biểu tượng',
-    'download_cv.free_title': 'Tải CV miễn phí',
-    'download_cv.free_description': 'Kèm biểu tượng @VCareer',
-      'download_cv.free_button': 'Tải CV miễn phí',
-      'download_cv.success_without_logo': 'Tải CV không kèm biểu tượng thành công!',
-      'download_cv.success_free': 'Tải CV miễn phí thành công!',
-    'cv_management.copy_link': 'Sao chép liên kết',
-    'cv_management.share_facebook': 'Chia sẻ trên Facebook',
-    'cv_management.create_copy': 'Tạo bản sao',
-      'cv_management.rename': 'Đổi tên',
-      'cv_management.rename_success': 'Đổi tên CV thành công!',
-      'cv_management.delete_success': 'Xóa CV thành công!',
-      'cv_management.upload_cv_title': 'CV đã tải lên VCareer',
-      'cv_management.upload_cv': 'Tải CV lên',
-      'cv_management.no_cv_uploaded': 'Chưa có CV nào được tải lên.',
-      'cv_management.uploading_cv': 'Đang tải CV lên...',
-      
-      // Upload CV Modal
-      'upload_cv.title': 'Upload CV để các cơ hội việc làm tự tìm đến bạn',
-      'upload_cv.subtitle': 'Giảm đến 50% thời gian cần thiết để tìm được một công việc phù hợp',
-      'upload_cv.description': 'Bạn đã có sẵn CV của mình, chỉ cần tải CV lên, hệ thống sẽ tự động đề xuất CV của bạn tới những nhà tuyển dụng uy tín.',
-      'upload_cv.benefits': 'Tiết kiệm thời gian, tìm việc thông minh, nắm bắt cơ hội và làm chủ đường đua nghề nghiệp của chính mình.',
-      'upload_cv.upload_instructions': 'Tải lên CV từ máy tính, chọn hoặc kéo thả',
-      'upload_cv.supported_formats': 'Hỗ trợ định dạng .doc, .docx, .pdf có kích thước dưới 5MB',
-      'upload_cv.select_cv': 'Chọn CV',
-      'upload_cv.upload_button': 'Tải CV lên',
-      'upload_cv.file_selected_success': 'File đã được chọn thành công!',
-      'upload_cv.invalid_file_type': 'Định dạng file không được hỗ trợ. Vui lòng chọn file .doc, .docx hoặc .pdf',
-      'upload_cv.file_too_large': 'File quá lớn. Vui lòng chọn file dưới 5MB',
-      'upload_cv.upload_success': 'Upload CV thành công!',
-      
-      // CV Card
-      'cv_card.updated': 'Cập nhật',
-      'cv_card.download': 'Tải về',
-      'cv_card.download_success': 'Tải CV thành công!',
-      'cv_card.copy_link': 'Sao chép liên kết',
-      'cv_card.share_facebook': 'Chia sẻ trên Facebook',
-      'cv_card.rename': 'Đổi tên',
-      'cv_card.delete': 'Xoá',
-      'cv_card.copy_link_success': 'Đã sao chép liên kết!',
-      'cv_card.share_facebook_success': 'Đã chia sẻ trên Facebook!',
-      'cv_card.rename_success': 'Đã đổi tên CV!',
-      'cv_card.delete_success': 'Đã xóa CV!',
-      
-      // Confirm Delete Modal
-      'confirm_delete.title': 'Xác nhận',
-      'confirm_delete.message': 'Bạn chắc chắn muốn xóa CV này?',
-      'confirm_delete.cancel': 'Hủy',
-      'confirm_delete.delete': 'Xóa',
-      
-      // Rename CV Modal
-      'rename_cv.title': 'Đổi tên CV',
-      'rename_cv.placeholder': 'Nhập tên CV mới',
-      'rename_cv.hint': 'Tên CV (Ví dụ: CV Marketing, CV Lập trình, CV ứng tuyển công ty...)',
-      'rename_cv.cancel': 'Hủy bỏ',
-      'rename_cv.update': 'Cập nhật',
-    
-    // Profile Sidebar
-    'profile_sidebar.welcome_back': 'Chào bạn trở lại',
-    'profile_sidebar.verified_account': 'Tài khoản đã xác thực',
-    'profile_sidebar.upgrade_account': 'Nâng cấp tài khoản',
-    'profile_sidebar.contact_description': 'Khi có cơ hội việc làm phù hợp, NTD sẽ liên hệ và trao đổi với bạn qua:',
-    'profile_sidebar.email_phone': 'Email và Số điện thoại của bạn',
-
-    // Profile Picture Edit Modal
-    'profile_picture_edit.title': 'CHỈNH SỬA ẢNH ĐẠI DIỆN',
-    'profile_picture_edit.preview_label': 'Ảnh hiển thị trên CV',
-    'profile_picture_edit.change_picture': 'Đổi ảnh',
-    'profile_picture_edit.delete_picture': 'Xóa ảnh',
-    'profile_picture_edit.done': 'Xong',
-    'profile_picture_edit.close_no_save': 'Đóng lại (Không lưu)',
-    'profile_picture_edit.change_success': 'Ảnh đại diện đã được thay đổi thành công!',
-    'profile_picture_edit.delete_success': 'Ảnh đại diện đã được xóa thành công!',
-      
       // Recruiter Header
       'recruiter.about': 'Giới thiệu',
       'recruiter.services': 'Dịch vụ',
@@ -394,9 +293,45 @@ export class TranslationService {
       'recruiter.blog': 'Blog tuyển dụng',
       'recruiter.post_job': 'Đăng tin ngay',
       
+      // Service Page
+      'service.title': 'Dịch vụ đăng tin tuyển dụng',
+      'service.description': 'Các dịch vụ tuyển dụng chuyên nghiệp của VCareer',
+      'service.free_posting_title': 'Đăng tin tuyển dụng miễn phí',
+      'service.free_posting_desc1': 'Đăng tin tuyển dụng miễn phí và không giới hạn số lượng.',
+      'service.free_posting_desc2': 'Đăng tin tuyển dụng dễ dàng, không quá 1 phút.',
+      'service.free_posting_desc3': 'Tiếp cận nguồn CV ứng viên khổng lồ, tìm kiếm ứng viên từ kho dữ liệu hơn 5 triệu hồ sơ.',
+      'service.free_posting_desc4': 'Dễ dàng kiểm duyệt và đăng tin trong 24h.',
+      'service.top_jobs_title': 'Top Jobs & Top Standard - Đăng tin tuyển dụng',
+      'service.top_jobs_desc1': 'Tăng lượt tiếp cận người tìm việc thêm 300% khi đăng tuyển.',
+      'service.top_jobs_desc2': 'Tin tuyển dụng hiển thị ở những vị trí nổi bật.',
+      'service.top_jobs_desc3': 'Đẩy tin tuyển dụng lên vị trí đầu trong kết quả tìm kiếm việc làm trên trang web đăng tin tuyển dụng.',
+      'service.top_jobs_desc4': 'Tự động gợi ý tin tuyển dụng với ứng viên phù hợp, giúp tuyển dụng hiệu quả hơn.',
+      'service.cv_suggest_title': 'CV đề xuất',
+      'service.cv_suggest_desc1': 'Đa dạng hóa nguồn CV ứng viên mà không cần mất công tìm kiếm ứng viên.',
+      'service.cv_suggest_desc2': 'Tiết kiệm thời gian tuyển dụng nhân sự.',
+      'service.cv_suggest_desc3': 'Tỷ lệ ứng viên phù hợp lên đến 40%.',
+      'service.cv_suggest_desc4': 'Dịch vụ có cam kết CV đang tìm kiếm công việc.',
+      
       // Language
       'language.vietnamese': 'Tiếng Việt',
       'language.english': 'English',
+      
+      // About Us - Recruiter
+      'about_us.vcareer': 'VCareer',
+      'about_us.introduction_subtitle': 'Nền tảng tuyển dụng giúp doanh nghiệp bứt phá!',
+      'about_us.post_job': 'Đăng tin',
+      'about_us.about_title': 'Về chúng tôi',
+      'about_us.description1': 'VCareer là nền tảng công nghệ nhân sự (HR Tech) thế hệ mới tại Việt Nam, hướng tới mục tiêu kết nối doanh nghiệp và nhân tài một cách nhanh chóng, chính xác và hiệu quả nhất.',
+      'about_us.description2': 'Lấy công nghệ làm nền tảng, VCareer xây dựng hệ sinh thái tuyển dụng toàn diện xoay quanh ba trụ cột chính:',
+      'about_us.recruiter_title': 'VCareer Recruiter:',
+      'about_us.recruiter_desc': 'Nền tảng tuyển dụng thông minh, giúp doanh nghiệp dễ dàng đăng tin, tiếp cận và sàng lọc ứng viên tiềm năng.',
+      'about_us.profile_title': 'VCareer Profile:',
+      'about_us.profile_desc': 'Hệ thống hồ sơ ứng viên được chuẩn hóa, hỗ trợ nhà tuyển dụng đánh giá nhanh và khách quan năng lực từng cá nhân.',
+      'about_us.insight_title': 'VCareer Insight:',
+      'about_us.insight_desc': 'Giải pháp phân tích dữ liệu và hành vi ứng viên, giúp doanh nghiệp tối ưu chiến lược tuyển dụng và nâng cao hiệu suất nhân sự.',
+      'about_us.description3': 'Hiện nay, VCareer đã và đang đồng hành cùng hàng chục nghìn ứng viên và doanh nghiệp trên toàn quốc, tạo nên hàng trăm nghìn cơ hội việc làm được kết nối mỗi năm.',
+      'about_us.description4': 'Thông qua việc liên tục nghiên cứu và ứng dụng các công nghệ tiên tiến — đặc biệt là Trí tuệ nhân tạo (AI) và phân tích dữ liệu lớn (Big Data) — VCareer hướng tới mục tiêu mang lại trải nghiệm tuyển dụng – tìm việc hiệu quả, minh bạch và bền vững hơn cho cộng đồng lao động Việt Nam.',
+      'about_us.tagline': 'VCareer – Kết nối cơ hội, kiến tạo tương lai.',
       
       // Homepage
       'homepage.title': 'Tìm việc làm mơ ước của bạn',
@@ -454,7 +389,156 @@ export class TranslationService {
       'Công nghệ Thông tin': 'Công nghệ Thông tin',
       'Tài chính - Ngân hàng': 'Tài chính - Ngân hàng',
       'Bất động sản': 'Bất động sản',
-      'Kế toán - Kiểm toán': 'Kế toán - Kiểm toán'
+      'Kế toán - Kiểm toán': 'Kế toán - Kiểm toán',
+      
+      // CV Sample
+      'cv_sample.breadcrumb': 'Mẫu CV',
+      'cv_sample.page_title': 'Mẫu CV',
+      'cv_sample.filter.all': 'Tất cả',
+      'cv_sample.filter.simple': 'Đơn giản',
+      'cv_sample.filter.professional': 'Chuyên nghiệp',
+      'cv_sample.filter.modern': 'Hiện đại',
+      'cv_sample.use_template': 'Sử dụng mẫu',
+      'cv_sample.type.standard': 'CV Tiêu chuẩn',
+      'cv_sample.tag.simple': 'Đơn giản',
+      'cv_sample.no_results': 'Không tìm thấy mẫu CV phù hợp',
+      
+      // Write CV
+      'write_cv.unnamed_cv': 'CV chưa đặt tên',
+      'write_cv.preview': 'Xem trước',
+      'write_cv.save_cv': 'Lưu CV',
+      'write_cv.edit_photo': 'Sửa ảnh',
+      'write_cv.fullname_placeholder': 'Họ và tên',
+      'write_cv.position_placeholder': 'Vị trí ứng tuyển',
+      'write_cv.birthdate_label': 'Ngày sinh',
+      'write_cv.birthdate_placeholder': 'DD/MM/YYYY',
+      'write_cv.gender_label': 'Giới tính',
+      'write_cv.gender_placeholder': 'Nam/Nữ/Khác',
+      'write_cv.phone_label': 'Điện thoại',
+      'write_cv.phone_placeholder': 'Nhập số điện thoại',
+      'write_cv.email_label': 'Email',
+      'write_cv.email_placeholder': 'Nhập email',
+      'write_cv.website_label': 'Website',
+      'write_cv.website_placeholder': 'Nhập website',
+      'write_cv.address_label': 'Địa chỉ',
+      'write_cv.address_placeholder': 'Nhập địa chỉ',
+      'write_cv.section_career_objective': 'Mục tiêu nghề nghiệp',
+      'write_cv.section_education': 'Học vấn',
+      'write_cv.start_date': 'Từ',
+      'write_cv.end_date': 'Đến',
+      'write_cv.section_work_experience': 'Kinh nghiệm làm việc',
+      'write_cv.section_activities': 'Hoạt động',
+      'write_cv.section_certificates': 'Chứng chỉ',
+      'write_cv.section_awards': 'Giải thưởng',
+      'write_cv.section_skills': 'Kỹ năng',
+      'write_cv.section_references': 'Người tham khảo',
+      'write_cv.section_hobbies': 'Sở thích',
+      'write_cv.career_objective_placeholder': 'Nhập mục tiêu nghề nghiệp của bạn',
+      'write_cv.institution_placeholder': 'Tên trường/đại học',
+      'write_cv.major_placeholder': 'Chuyên ngành',
+      'write_cv.education_description': 'Mô tả thêm về học vấn',
+      'write_cv.company_placeholder': 'Tên công ty',
+      'write_cv.job_title_placeholder': 'Chức vụ',
+      'write_cv.work_description': 'Mô tả công việc',
+      'write_cv.activity_organization_placeholder': 'Tên tổ chức',
+      'write_cv.activity_position_placeholder': 'Vị trí',
+      'write_cv.activity_description': 'Mô tả hoạt động',
+      'write_cv.certificate_time': 'Thời gian',
+      'write_cv.certificate_name_placeholder': 'Tên chứng chỉ',
+      'write_cv.award_time': 'Thời gian',
+      'write_cv.award_name_placeholder': 'Tên giải thưởng',
+      'write_cv.skill_name_placeholder': 'Tên kỹ năng',
+      'write_cv.skill_description_placeholder': 'Mô tả kỹ năng',
+      'write_cv.reference_placeholder': 'Thông tin người tham khảo',
+      'write_cv.hobby_placeholder': 'Sở thích của bạn',
+      'write_cv.preview_dialog_title': 'Xem trước CV',
+      'write_cv.modal_title': 'Đặt tên cho CV',
+      'write_cv.modal_info': 'Vui lòng đặt tên cho CV của bạn',
+      'write_cv.modal_input_placeholder': 'Nhập tên CV',
+      'write_cv.modal_back_button': 'Hủy',
+      'write_cv.modal_continue_button': 'Tiếp tục',
+      'write_cv.save_success': 'Lưu CV thành công',
+      'write_cv.save_failed': 'Lưu CV thất bại',
+      'write_cv.validation_message': 'Vui lòng điền đầy đủ thông tin bắt buộc',
+      'write_cv.photo_modal_title': 'Cập nhật ảnh đại diện',
+      'write_cv.photo_modal_instruction': 'Tải lên ảnh đại diện của bạn',
+      'write_cv.photo_modal_note': 'Hỗ trợ định dạng JPG, PNG. Kích thước tối đa 5MB',
+      'write_cv.photo_modal_cancel': 'Hủy',
+      'write_cv.photo_modal_confirm': 'Xác nhận',
+      'write_cv.photo_update_success': 'Cập nhật ảnh thành công',
+      
+      // CV Management
+      'cv_management.title': 'Quản lý CV',
+      'cv_management.subtitle': 'Tạo, chỉnh sửa và quản lý CV của bạn',
+      'cv_management.cv_list_title': 'Danh sách CV',
+      'cv_management.no_cv_title': 'Chưa có CV nào',
+      'cv_management.no_cv_description': 'Bắt đầu tạo CV đầu tiên của bạn để ứng tuyển việc làm',
+      'cv_management.upload_cv_title': 'CV đã tải lên',
+      'cv_management.no_cv_uploaded': 'Chưa có CV nào được tải lên',
+      'cv_management.upload_cv': 'Tải lên CV',
+      'cv_management.create_cv': 'Tạo CV',
+      'cv_management.create_first_cv': 'Tạo CV đầu tiên',
+      'cv_management.download': 'Tải xuống',
+      'cv_management.edit': 'Chỉnh sửa',
+      'cv_management.push_to_top': 'Đẩy lên đầu',
+      'cv_management.copy_link': 'Sao chép liên kết',
+      'cv_management.share_facebook': 'Chia sẻ Facebook',
+      'cv_management.create_copy': 'Tạo bản sao',
+      'cv_management.rename': 'Đổi tên',
+      'cv_management.delete': 'Xóa',
+      'cv_management.updated': 'Cập nhật',
+      'cv_management.loading': 'Đang tải...',
+      'cv_management.updated_successfully': 'Cập nhật CV thành công',
+      'cv_management.deleted_successfully': 'Xóa CV thành công',
+      'cv_management.duplicated_successfully': 'Sao chép CV thành công',
+      'cv_management.set_default_successfully': 'Đặt CV mặc định thành công',
+      
+      // CV Card
+      'cv_card.updated': 'Cập nhật',
+      'cv_card.copy_link_success': 'Đã sao chép liên kết',
+      'cv_card.delete_success': 'Đã xóa CV',
+      'cv_card.share_facebook_success': 'Đã chia sẻ lên Facebook',
+      'cv_card.rename_success': 'Đã đổi tên CV',
+      
+      // Profile Sidebar
+      'profile_sidebar.welcome_back': 'Chào mừng trở lại',
+      'profile_sidebar.verified_account': 'Tài khoản đã xác thực',
+      'profile_sidebar.upgrade_account': 'Nâng cấp tài khoản',
+      'profile_sidebar.contact_description': 'Nhà tuyển dụng có thể liên hệ với bạn qua',
+      'profile_sidebar.email_phone': 'Email & Số điện thoại',
+      
+      // Profile Picture Edit
+      'profile_picture_edit.title': 'Chỉnh sửa ảnh đại diện',
+      'profile_picture_edit.preview_label': 'Xem trước',
+      'profile_picture_edit.change_picture': 'Thay đổi ảnh',
+      'profile_picture_edit.delete_picture': 'Xóa ảnh',
+      'profile_picture_edit.done': 'Hoàn thành',
+      'profile_picture_edit.close_no_save': 'Đóng không lưu',
+      'profile_picture_edit.change_success': 'Đã thay đổi ảnh đại diện',
+      'profile_picture_edit.delete_success': 'Đã xóa ảnh đại diện',
+      
+      // Upload CV
+      'upload_cv.upload_success': 'Tải lên CV thành công',
+      
+      // Download CV
+      'download_cv.success_without_logo': 'Tải xuống CV thành công (không logo)',
+      'download_cv.success_free': 'Tải xuống CV miễn phí thành công',
+      
+      // Saved Jobs
+      'saved_jobs.title': 'Việc làm đã lưu',
+      'saved_jobs.description': 'Xem lại danh sách những việc làm mà bạn đã lưu trước đó. Ứng tuyển ngay để không bỏ lỡ cơ hội nghề nghiệp dành cho bạn.',
+      'saved_jobs.list_count': 'Danh sách {{count}} việc làm đã lưu',
+      'saved_jobs.list_count_one': 'Danh sách 1 việc làm đã lưu',
+      'saved_jobs.no_jobs_count': 'Chưa có việc làm đã lưu',
+      'saved_jobs.saved_date': 'Đã lưu',
+      'saved_jobs.updated': 'Cập nhật',
+      'saved_jobs.apply': 'Ứng tuyển',
+      'saved_jobs.unsave': 'Bỏ lưu',
+      'saved_jobs.unsave_success': 'Đã bỏ lưu việc làm',
+      'saved_jobs.no_jobs_title': 'Chưa có việc làm đã lưu',
+      'saved_jobs.no_jobs_description': 'Bạn chưa lưu việc làm nào. Hãy duyệt và lưu những việc làm phù hợp với bạn.',
+      'saved_jobs.browse_jobs': 'Duyệt việc làm',
+      'saved_jobs.loading': 'Đang tải...'
     },
     en: {
       // Breadcrumb
@@ -495,6 +579,36 @@ export class TranslationService {
       'profile.upgrade_account': 'Upgrade account',
       'profile.upgrade_vip_account': 'Upgrade VIP account',
       'profile.activate_gift': 'Activate gift',
+      
+      // Notifications
+      'notifications.title': 'Notifications',
+      'notifications.mark_all_read': 'Mark all as read',
+      'notifications.no_notifications': 'No new notifications',
+      
+      // Sidebar
+      'sidebar.verification': 'Account verification',
+      'sidebar.not_verified': 'Not verified',
+      'sidebar.dashboard': 'Dashboard',
+      'sidebar.insights': 'TopCV Insights',
+      'sidebar.rewards': 'TopCV Rewards',
+      'sidebar.redeem_gifts': 'Redeem Gifts',
+      'sidebar.toppy_ai': 'Toppy AI - Suggestions',
+      'sidebar.suggested_cv': 'Suggested CVs',
+      'sidebar.recruitment_campaigns': 'Recruitment Campaigns',
+      'sidebar.job_postings': 'Job Postings',
+      'sidebar.cv_management': 'CV Management',
+      'sidebar.recruitment_reports': 'Recruitment Reports',
+      'sidebar.buy_services': 'Buy Services',
+      'sidebar.account_settings': 'Account Settings',
+      'sidebar.activity_history': 'Activity History',
+      
+      // Verification
+      'verify.step_phone': 'Verify phone number',
+      'verify.step_company': 'Update company information',
+      'verify.step_license': 'Update business registration certificate',
+      'verify.step_first_job': 'Post first job',
+      'verify.next_step_phone': 'Verify phone number',
+      'verify.complete': 'Complete',
       
       // Job Page
       'job_page.title': 'Jobs',
@@ -699,137 +813,6 @@ export class TranslationService {
       
       'profile.logout': 'Logout',
       
-      // Notifications
-      'notifications.title': 'Notifications',
-      'notifications.mark_all_read': 'Mark as read',
-      'notifications.sample_text': 'Candidates in your industry are also interested in this position:',
-
-      // CV Management
-      'cv_management.title': 'CV Management',
-      'cv_management.subtitle': 'Create and manage your CV easily',
-      'cv_management.cv_list_title': 'CVs created on VCareer',
-      'cv_management.create_cv': 'Create CV',
-      'cv_management.create_first_cv': 'Create first CV',
-      'cv_management.no_cv_title': 'No CVs yet',
-      'cv_management.no_cv_description': 'Create your first CV to start job searching',
-      'cv_management.updated': 'Updated',
-      'cv_management.view': 'View',
-      'cv_management.edit': 'Edit',
-      'cv_management.duplicate': 'Duplicate',
-      'cv_management.set_default': 'Set as default',
-      'cv_management.delete': 'Delete',
-      'cv_management.confirm_delete': 'Are you sure you want to delete this CV?',
-      'cv_management.welcome_back': 'Welcome back',
-      'cv_management.verified_account': 'Verified account',
-      'cv_management.upgrade_account': 'Upgrade account',
-      'cv_management.job_search_settings': 'Job search settings',
-      'cv_management.job_search_status': 'Job search is OFF',
-      'cv_management.job_search_description': 'Enable job search to make your profile more prominent and get more attention from recruiters.',
-      'cv_management.allow_recruiter_search': 'Allow recruiters to search profile',
-      'cv_management.contact_methods_title': 'When there are suitable job opportunities, recruiters will contact you via:',
-      'cv_management.top_connect': 'Top Connect messaging on VCareer',
-      'cv_management.email_phone': 'Your email and phone number',
-      'cv_management.download_app': 'Download VCareer App now!',
-      'cv_management.app_description': 'Don\'t miss any opportunities from recruiters',
-      'cv_management.career_guidance': 'Career guidance by astrology',
-      'cv_management.career_feature_1': 'Understand your career overview',
-      'cv_management.career_feature_2': 'Career guidance by astrology',
-      'cv_management.career_feature_3': 'Master your destiny with experts',
-      'cv_management.register_now': 'Register now',
-      'cv_management.hide_profile': 'Hide profile from recruiters',
-      'cv_management.new': 'New',
-      'cv_management.hide_profile_description': 'I don\'t want my CV to be displayed to recruiters with email domains and companies listed below:',
-      'cv_management.default': 'Default',
-      'cv_management.loading': 'Loading...',
-      'cv_management.creating_cv': 'Creating CV...',
-      'cv_management.updated_successfully': 'CV updated successfully!',
-      'cv_management.deleted_successfully': 'CV deleted successfully!',
-      'cv_management.delete_failed': 'Failed to delete CV',
-      'cv_management.duplicated_successfully': 'CV duplicated successfully!',
-      'cv_management.duplicate_failed': 'Failed to duplicate CV',
-      'cv_management.set_default_successfully': 'CV set as default successfully!',
-      'cv_management.set_default_failed': 'Failed to set CV as default',
-      'cv_management.download': 'Download',
-      'cv_management.push_to_top': 'Push to top',
-
-      // Download CV Modal
-      'download_cv.title': 'Download CV',
-      'download_cv.without_logo_title': 'Download CV without @VCareer logo',
-      'download_cv.without_logo_description': 'Unlimited downloads, CVs and design templates within 24 hours.',
-      'download_cv.without_logo_button': 'Download CV without logo',
-      'download_cv.free_title': 'Free download',
-      'download_cv.free_description': 'With @VCareer logo',
-      'download_cv.free_button': 'Free download',
-      'download_cv.success_without_logo': 'Download CV without logo successful!',
-      'download_cv.success_free': 'Free download successful!',
-      'cv_management.copy_link': 'Copy link',
-      'cv_management.share_facebook': 'Share on Facebook',
-      'cv_management.create_copy': 'Create copy',
-      'cv_management.rename': 'Rename',
-      'cv_management.rename_success': 'CV renamed successfully!',
-      'cv_management.delete_success': 'CV deleted successfully!',
-      'cv_management.upload_cv_title': 'CVs uploaded to VCareer',
-      'cv_management.upload_cv': 'Upload CV',
-      'cv_management.no_cv_uploaded': 'No CVs have been uploaded yet.',
-      'cv_management.uploading_cv': 'Uploading CV...',
-      
-      // Upload CV Modal
-      'upload_cv.title': 'Upload CV so job opportunities find you',
-      'upload_cv.subtitle': 'Reduce up to 50% of the time needed to find a suitable job',
-      'upload_cv.description': 'If you already have your CV, just upload it, and the system will automatically suggest your CV to reputable recruiters.',
-      'upload_cv.benefits': 'Save time, find jobs smartly, seize opportunities, and take control of your career path.',
-      'upload_cv.upload_instructions': 'Upload CV from computer, select or drag and drop',
-      'upload_cv.supported_formats': 'Supports .doc, .docx, .pdf formats with size under 5MB',
-      'upload_cv.select_cv': 'Select CV',
-      'upload_cv.upload_button': 'Upload CV',
-      'upload_cv.file_selected_success': 'File selected successfully!',
-      'upload_cv.invalid_file_type': 'Unsupported file format. Please select .doc, .docx or .pdf file',
-      'upload_cv.file_too_large': 'File too large. Please select file under 5MB',
-      'upload_cv.upload_success': 'CV uploaded successfully!',
-      
-      // CV Card
-      'cv_card.updated': 'Updated',
-      'cv_card.download': 'Download',
-      'cv_card.download_success': 'CV downloaded successfully!',
-      'cv_card.copy_link': 'Copy link',
-      'cv_card.share_facebook': 'Share on Facebook',
-      'cv_card.rename': 'Rename',
-      'cv_card.delete': 'Delete',
-      'cv_card.copy_link_success': 'Link copied!',
-      'cv_card.share_facebook_success': 'Shared on Facebook!',
-      'cv_card.rename_success': 'CV renamed!',
-      'cv_card.delete_success': 'CV deleted!',
-      
-      // Confirm Delete Modal
-      'confirm_delete.title': 'Confirm',
-      'confirm_delete.message': 'Are you sure you want to delete this CV?',
-      'confirm_delete.cancel': 'Cancel',
-      'confirm_delete.delete': 'Delete',
-      
-      // Rename CV Modal
-      'rename_cv.title': 'Rename CV',
-      'rename_cv.placeholder': 'Enter new CV name',
-      'rename_cv.hint': 'CV Name (Example: CV Marketing, CV Programming, CV company application...)',
-      'rename_cv.cancel': 'Cancel',
-      'rename_cv.update': 'Update',
-      
-      // Profile Sidebar
-      'profile_sidebar.welcome_back': 'Welcome back',
-      'profile_sidebar.verified_account': 'Verified account',
-      'profile_sidebar.upgrade_account': 'Upgrade account',
-      'profile_sidebar.contact_description': 'When there is a suitable job opportunity, the Employer will contact and discuss with you via:',
-      'profile_sidebar.email_phone': 'Your Email and Phone number',
-
-      // Profile Picture Edit Modal
-      'profile_picture_edit.title': 'EDIT PROFILE PICTURE',
-      'profile_picture_edit.preview_label': 'Picture displayed on CV',
-      'profile_picture_edit.change_picture': 'Change picture',
-      'profile_picture_edit.delete_picture': 'Delete picture',
-      'profile_picture_edit.done': 'Done',
-      'profile_picture_edit.close_no_save': 'Close (Don\'t save)',
-      'profile_picture_edit.change_success': 'Profile picture changed successfully!',
-      'profile_picture_edit.delete_success': 'Profile picture deleted successfully!',
-      
       // Recruiter Header
       'recruiter.about': 'About',
       'recruiter.services': 'Services',
@@ -838,9 +821,45 @@ export class TranslationService {
       'recruiter.blog': 'Recruitment Blog',
       'recruiter.post_job': 'Post Job Now',
       
+      // Service Page
+      'service.title': 'Job Posting Services',
+      'service.description': 'Professional recruitment services from VCareer',
+      'service.free_posting_title': 'Free Job Posting',
+      'service.free_posting_desc1': 'Free job posting with unlimited number of posts.',
+      'service.free_posting_desc2': 'Easy job posting, less than 1 minute.',
+      'service.free_posting_desc3': 'Access to a huge database of candidate CVs, search from over 5 million profiles.',
+      'service.free_posting_desc4': 'Easy review and post within 24 hours.',
+      'service.top_jobs_title': 'Top Jobs & Top Standard - Job Posting',
+      'service.top_jobs_desc1': 'Increase job seeker reach by 300% when posting.',
+      'service.top_jobs_desc2': 'Job postings displayed in prominent positions.',
+      'service.top_jobs_desc3': 'Push job postings to the top of search results on the job posting website.',
+      'service.top_jobs_desc4': 'Automatically suggest job postings to suitable candidates, helping with more effective recruitment.',
+      'service.cv_suggest_title': 'CV Recommendations',
+      'service.cv_suggest_desc1': 'Diversify candidate CV sources without the need to search for candidates.',
+      'service.cv_suggest_desc2': 'Save time in recruitment.',
+      'service.cv_suggest_desc3': 'Up to 40% of candidates match the requirements.',
+      'service.cv_suggest_desc4': 'Service guarantees CVs from candidates actively seeking jobs.',
+      
       // Language
       'language.vietnamese': 'Tiếng Việt',
       'language.english': 'English',
+      
+      // About Us - Recruiter
+      'about_us.vcareer': 'VCareer',
+      'about_us.introduction_subtitle': 'Recruitment platform helping businesses break through!',
+      'about_us.post_job': 'Post Job',
+      'about_us.about_title': 'About Us',
+      'about_us.description1': 'VCareer is a new-generation HR Tech platform in Vietnam, aimed at connecting businesses and talent in the fastest, most accurate, and most effective way.',
+      'about_us.description2': 'With technology as the foundation, VCareer builds a comprehensive recruitment ecosystem around three main pillars:',
+      'about_us.recruiter_title': 'VCareer Recruiter:',
+      'about_us.recruiter_desc': 'An intelligent recruitment platform that helps businesses easily post jobs, reach out to and screen potential candidates.',
+      'about_us.profile_title': 'VCareer Profile:',
+      'about_us.profile_desc': 'A standardized candidate profile system that helps recruiters quickly and objectively assess each individual\'s capabilities.',
+      'about_us.insight_title': 'VCareer Insight:',
+      'about_us.insight_desc': 'A data analytics and candidate behavior analysis solution that helps businesses optimize recruitment strategies and improve HR performance.',
+      'about_us.description3': 'Currently, VCareer has been and continues to accompany tens of thousands of candidates and businesses nationwide, creating hundreds of thousands of job opportunities connected each year.',
+      'about_us.description4': 'Through continuous research and application of advanced technologies — especially Artificial Intelligence (AI) and Big Data analytics — VCareer aims to deliver more effective, transparent, and sustainable recruitment and job-seeking experiences for Vietnam\'s workforce community.',
+      'about_us.tagline': 'VCareer – Connecting opportunities, creating the future.',
       
       // Homepage
       'homepage.title': 'Find Your Dream Job',
@@ -898,7 +917,156 @@ export class TranslationService {
       'Công nghệ Thông tin': 'Information Technology',
       'Tài chính - Ngân hàng': 'Finance - Banking',
       'Bất động sản': 'Real Estate',
-      'Kế toán - Kiểm toán': 'Accounting - Auditing'
+      'Kế toán - Kiểm toán': 'Accounting - Auditing',
+      
+      // CV Sample
+      'cv_sample.breadcrumb': 'CV Samples',
+      'cv_sample.page_title': 'CV Samples',
+      'cv_sample.filter.all': 'All',
+      'cv_sample.filter.simple': 'Simple',
+      'cv_sample.filter.professional': 'Professional',
+      'cv_sample.filter.modern': 'Modern',
+      'cv_sample.use_template': 'Use Template',
+      'cv_sample.type.standard': 'Standard CV',
+      'cv_sample.tag.simple': 'Simple',
+      'cv_sample.no_results': 'No matching CV samples found',
+      
+      // Write CV
+      'write_cv.unnamed_cv': 'Unnamed CV',
+      'write_cv.preview': 'Preview',
+      'write_cv.save_cv': 'Save CV',
+      'write_cv.edit_photo': 'Edit Photo',
+      'write_cv.fullname_placeholder': 'Full Name',
+      'write_cv.position_placeholder': 'Applied Position',
+      'write_cv.birthdate_label': 'Date of Birth',
+      'write_cv.birthdate_placeholder': 'DD/MM/YYYY',
+      'write_cv.gender_label': 'Gender',
+      'write_cv.gender_placeholder': 'Male/Female/Other',
+      'write_cv.phone_label': 'Phone',
+      'write_cv.phone_placeholder': 'Enter phone number',
+      'write_cv.email_label': 'Email',
+      'write_cv.email_placeholder': 'Enter email',
+      'write_cv.website_label': 'Website',
+      'write_cv.website_placeholder': 'Enter website',
+      'write_cv.address_label': 'Address',
+      'write_cv.address_placeholder': 'Enter address',
+      'write_cv.section_career_objective': 'Career Objective',
+      'write_cv.section_education': 'Education',
+      'write_cv.start_date': 'From',
+      'write_cv.end_date': 'To',
+      'write_cv.section_work_experience': 'Work Experience',
+      'write_cv.section_activities': 'Activities',
+      'write_cv.section_certificates': 'Certificates',
+      'write_cv.section_awards': 'Awards',
+      'write_cv.section_skills': 'Skills',
+      'write_cv.section_references': 'References',
+      'write_cv.section_hobbies': 'Hobbies',
+      'write_cv.career_objective_placeholder': 'Enter your career objective',
+      'write_cv.institution_placeholder': 'School/University name',
+      'write_cv.major_placeholder': 'Major',
+      'write_cv.education_description': 'Additional education description',
+      'write_cv.company_placeholder': 'Company name',
+      'write_cv.job_title_placeholder': 'Job Title',
+      'write_cv.work_description': 'Job description',
+      'write_cv.activity_organization_placeholder': 'Organization name',
+      'write_cv.activity_position_placeholder': 'Position',
+      'write_cv.activity_description': 'Activity description',
+      'write_cv.certificate_time': 'Time',
+      'write_cv.certificate_name_placeholder': 'Certificate name',
+      'write_cv.award_time': 'Time',
+      'write_cv.award_name_placeholder': 'Award name',
+      'write_cv.skill_name_placeholder': 'Skill name',
+      'write_cv.skill_description_placeholder': 'Skill description',
+      'write_cv.reference_placeholder': 'Reference information',
+      'write_cv.hobby_placeholder': 'Your hobbies',
+      'write_cv.preview_dialog_title': 'Preview CV',
+      'write_cv.modal_title': 'Name Your CV',
+      'write_cv.modal_info': 'Please name your CV',
+      'write_cv.modal_input_placeholder': 'Enter CV name',
+      'write_cv.modal_back_button': 'Cancel',
+      'write_cv.modal_continue_button': 'Continue',
+      'write_cv.save_success': 'CV saved successfully',
+      'write_cv.save_failed': 'Failed to save CV',
+      'write_cv.validation_message': 'Please fill in all required information',
+      'write_cv.photo_modal_title': 'Update Profile Photo',
+      'write_cv.photo_modal_instruction': 'Upload your profile photo',
+      'write_cv.photo_modal_note': 'Supports JPG, PNG formats. Maximum size 5MB',
+      'write_cv.photo_modal_cancel': 'Cancel',
+      'write_cv.photo_modal_confirm': 'Confirm',
+      'write_cv.photo_update_success': 'Photo updated successfully',
+      
+      // CV Management
+      'cv_management.title': 'CV Management',
+      'cv_management.subtitle': 'Create, edit and manage your CVs',
+      'cv_management.cv_list_title': 'CV List',
+      'cv_management.no_cv_title': 'No CV yet',
+      'cv_management.no_cv_description': 'Start creating your first CV to apply for jobs',
+      'cv_management.upload_cv_title': 'Uploaded CVs',
+      'cv_management.no_cv_uploaded': 'No CV uploaded yet',
+      'cv_management.upload_cv': 'Upload CV',
+      'cv_management.create_cv': 'Create CV',
+      'cv_management.create_first_cv': 'Create First CV',
+      'cv_management.download': 'Download',
+      'cv_management.edit': 'Edit',
+      'cv_management.push_to_top': 'Push to Top',
+      'cv_management.copy_link': 'Copy Link',
+      'cv_management.share_facebook': 'Share on Facebook',
+      'cv_management.create_copy': 'Create Copy',
+      'cv_management.rename': 'Rename',
+      'cv_management.delete': 'Delete',
+      'cv_management.updated': 'Updated',
+      'cv_management.loading': 'Loading...',
+      'cv_management.updated_successfully': 'CV updated successfully',
+      'cv_management.deleted_successfully': 'CV deleted successfully',
+      'cv_management.duplicated_successfully': 'CV duplicated successfully',
+      'cv_management.set_default_successfully': 'CV set as default successfully',
+      
+      // CV Card
+      'cv_card.updated': 'Updated',
+      'cv_card.copy_link_success': 'Link copied',
+      'cv_card.delete_success': 'CV deleted',
+      'cv_card.share_facebook_success': 'Shared on Facebook',
+      'cv_card.rename_success': 'CV renamed',
+      
+      // Profile Sidebar
+      'profile_sidebar.welcome_back': 'Welcome back',
+      'profile_sidebar.verified_account': 'Verified Account',
+      'profile_sidebar.upgrade_account': 'Upgrade Account',
+      'profile_sidebar.contact_description': 'Recruiters can contact you via',
+      'profile_sidebar.email_phone': 'Email & Phone',
+      
+      // Profile Picture Edit
+      'profile_picture_edit.title': 'Edit Profile Picture',
+      'profile_picture_edit.preview_label': 'Preview',
+      'profile_picture_edit.change_picture': 'Change Picture',
+      'profile_picture_edit.delete_picture': 'Delete Picture',
+      'profile_picture_edit.done': 'Done',
+      'profile_picture_edit.close_no_save': 'Close without saving',
+      'profile_picture_edit.change_success': 'Profile picture changed',
+      'profile_picture_edit.delete_success': 'Profile picture deleted',
+      
+      // Upload CV
+      'upload_cv.upload_success': 'CV uploaded successfully',
+      
+      // Download CV
+      'download_cv.success_without_logo': 'CV downloaded successfully (without logo)',
+      'download_cv.success_free': 'CV downloaded successfully (free)',
+      
+      // Saved Jobs
+      'saved_jobs.title': 'Saved Jobs',
+      'saved_jobs.description': 'Review the list of jobs you have saved previously. Apply now to not miss out on career opportunities for you.',
+      'saved_jobs.list_count': 'List of {{count}} saved jobs',
+      'saved_jobs.list_count_one': 'List of 1 saved job',
+      'saved_jobs.no_jobs_count': 'No saved jobs yet',
+      'saved_jobs.saved_date': 'Saved',
+      'saved_jobs.updated': 'Updated',
+      'saved_jobs.apply': 'Apply',
+      'saved_jobs.unsave': 'Unsave',
+      'saved_jobs.unsave_success': 'Job unsaved',
+      'saved_jobs.no_jobs_title': 'No saved jobs yet',
+      'saved_jobs.no_jobs_description': 'You haven\'t saved any jobs yet. Browse and save jobs that match your interests.',
+      'saved_jobs.browse_jobs': 'Browse Jobs',
+      'saved_jobs.loading': 'Loading...'
       ,
       // Category names (extra pages)
       'Sản xuất': 'Manufacturing',
