@@ -1,6 +1,10 @@
-import { Component } from '@angular/core';
+import { Component, OnInit, AfterViewInit, OnDestroy } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { Router, NavigationEnd } from '@angular/router';
 import { TranslationService } from '../../../../core/services/translation.service';
+import { NavigationService } from '../../../../core/services/navigation.service';
+import { filter } from 'rxjs/operators';
+import { Subscription } from 'rxjs';
 
 @Component({
   selector: 'app-about-us',
@@ -9,13 +13,63 @@ import { TranslationService } from '../../../../core/services/translation.servic
   templateUrl: './about-us.html',
   styleUrl: './about-us.scss'
 })
-export class AboutUs {
-  selectedLanguage: string = 'vi';
+export class AboutUs implements OnInit, AfterViewInit, OnDestroy {
+  imagePath: string = 'assets/images/about-us/introduction.png';
+  private routerSubscription?: Subscription;
 
-  constructor(private translationService: TranslationService) {
-    this.translationService.currentLanguage$.subscribe(lang => {
-      this.selectedLanguage = lang;
-    });
+  constructor(
+    private translationService: TranslationService,
+    private router: Router,
+    private navigationService: NavigationService
+  ) {}
+
+  ngOnInit() {
+    this.routerSubscription = this.router.events
+      .pipe(filter(event => event instanceof NavigationEnd))
+      .subscribe((event: NavigationEnd) => {
+        if (event.url === '/recruiter/service') {
+          setTimeout(() => this.scrollToService(), 300);
+        } else if (event.url === '/recruiter/about-us') {
+          setTimeout(() => window.scrollTo({ top: 0, behavior: 'smooth' }), 100);
+        }
+      });
+  }
+
+  ngAfterViewInit() {
+    if (this.router.url === '/recruiter/service') {
+      setTimeout(() => this.scrollToService(), 300);
+    } else if (this.router.url === '/recruiter/about-us') {
+      setTimeout(() => window.scrollTo({ top: 0, behavior: 'smooth' }), 100);
+    }
+  }
+
+  ngOnDestroy() {
+    if (this.routerSubscription) {
+      this.routerSubscription.unsubscribe();
+    }
+  }
+
+  scrollToService(): void {
+    const element = document.getElementById('service-section');
+    if (element) {
+      element.scrollIntoView({ behavior: 'smooth', block: 'start' });
+    }
+  }
+
+  onPostJob(): void {
+    if (!this.navigationService.isLoggedIn()) {
+      this.router.navigate(['/recruiter/login']);
+    } else {
+      // Route post-job chưa được implement
+      // Nếu đã verified, không làm gì (hoặc có thể navigate đến home khi route được implement)
+      // Nếu chưa verified, navigate đến verify page
+      const isVerified = this.navigationService.isVerified();
+      if (!isVerified) {
+        this.router.navigate(['/recruiter/recruiter-verify']);
+      }
+      // TODO: Navigate to /recruiter/post-job when route is implemented
+      // For now, if already verified, do nothing or show message
+    }
   }
 
   translate(key: string): string {
