@@ -1,14 +1,15 @@
+import * as Application from './application';
 import * as Books from './books';
-import * as Controllers from './controllers';
 import * as Cv from './cv';
 import * as Dto from './dto';
+import * as HttpApi from './http-api';
 import * as Iservices from './iservices';
 import * as Job from './job';
 import * as Microsoft from './microsoft';
-import * as Model from './models';
+import * as Model from './model';
 import * as Models from './models';
 import * as Profile from './profile';
 import * as Services from './services';
 import * as Volo from './volo';
 export * from './api-config.service';
-export { Books, Controllers, Cv, Dto, Iservices, Job, Microsoft, Model, Models, Profile, Services, Volo };
+export { Application, Books, Cv, Dto, HttpApi, Iservices, Job, Microsoft, Model, Models, Profile, Services, Volo };
