@@ -2,7 +2,7 @@
 export interface IFormFile {
   contentType?: string;
   contentDisposition?: string;
-  // headers: Record<string, StringValues>;
+  headers: Record<string, string | string[]>;
   length: number;
   name?: string;
   fileName?: string;
