@@ -1,11 +1,8 @@
 
-// export interface ActionResult {
-// }
-
-// export interface ActionResult<TValue> {
-//   result: ActionResult;
-//   value: TValue;
-// }
+export interface ActionResult<TValue = any> {
+  result?: ActionResult<TValue>;
+  value?: TValue;
+}
 
 export interface IActionResult {
 }
