@@ -9,24 +9,6 @@ export class FileServicesService {
   apiName = 'Default';
   
 
-  delete = (storageName: string, userId: string, config?: Partial<Rest.Config>) =>
-    this.restService.request<any, void>({
-      method: 'DELETE',
-      url: '/api/app/file-services',
-      params: { storageName, userId },
-    },
-    { apiName: this.apiName,...config });
-  
-
-  download = (storageName: string, config?: Partial<Rest.Config>) =>
-    this.restService.request<any, FileDescriptorDto>({
-      method: 'POST',
-      url: '/api/app/file-services/download',
-      params: { storageName },
-    },
-    { apiName: this.apiName,...config });
-  
-
   getMetadata = (storageName: string, config?: Partial<Rest.Config>) =>
     this.restService.request<any, FileDescriptorDto>({
       method: 'GET',
@@ -36,9 +18,26 @@ export class FileServicesService {
     { apiName: this.apiName,...config });
   
 
-  upload = (input: UploadFileDto, config?: Partial<Rest.Config>) =>
+  hardDelete = (fileId: string, config?: Partial<Rest.Config>) =>
     this.restService.request<any, void>({
       method: 'POST',
+      url: `/api/app/file-services/hard-delete/${fileId}`,
+    },
+    { apiName: this.apiName,...config });
+  
+
+  softDelete = (fileId: string, config?: Partial<Rest.Config>) =>
+    this.restService.request<any, void>({
+      method: 'POST',
+      url: `/api/app/file-services/soft-delete/${fileId}`,
+    },
+    { apiName: this.apiName,...config });
+  
+
+  upload = (input: UploadFileDto, config?: Partial<Rest.Config>) =>
+    this.restService.request<any, string>({
+      method: 'POST',
+      responseType: 'text',
       url: '/api/app/file-services/upload',
       body: input,
     },
