@@ -1,5 +1,5 @@
 using System;
-using VCareer.Constants;
+using System.Collections.Generic;
 using VCareer.Model;
 
 namespace VCareer.Dto.Job
@@ -110,6 +110,21 @@ namespace VCareer.Dto.Job
         /// Lượt apply
         /// </summary>
         public int ApplyCount { get; set; }
+
+        /// <summary>
+        /// Trình độ học vấn yêu cầu
+        /// </summary>
+        public EducationLevel Education { get; set; }
+
+        /// <summary>
+        /// Chuỗi danh mục (từ cấp 1 -> cấp 3) của job hiện tại
+        /// </summary>
+        public List<CategoryItemDto> CategoryPath { get; set; } = new List<CategoryItemDto>();
+
+        /// <summary>
+        /// Job đã được người dùng hiện tại lưu (favorite) hay chưa
+        /// </summary>
+        public bool IsSaved { get; set; } = false;
     }
 }
 
