@@ -11,10 +11,12 @@ namespace VCareer.IServices.IFileServices
 {
     public interface IFileServices :IApplicationService
     {
-        Task UploadAsync(UploadFileDto input);
-        Task<FileDescriptorDto> DownloadAsync(string storageName);
-        Task DeleteAsync(string storageName, Guid userId);
+        Task<Guid> UploadAsync(UploadFileDto input); // Return FileDescriptor Id after upload
+    /*    Task<FileDescriptorDto> DownloadAsync(string storageName);*/
+        public Task HardDeleteAsync(string fileId);
+        public Task SoftDeleteAsync(string fileId);
         Task<FileDescriptorDto> GetMetadataAsync(string storageName);
         Task<FileChunkResult> UploadChunkAsync(FileChunkInput input);
+        
     }
 }
