@@ -9,6 +9,7 @@ using Microsoft.AspNetCore.HttpOverrides;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
+using Microsoft.IdentityModel.Tokens;
 using Microsoft.OpenApi.Models;
 using OpenIddict.Server.AspNetCore;
 using OpenIddict.Validation.AspNetCore;
@@ -54,6 +55,8 @@ using Volo.Abp.UI.Navigation.Urls;
 using Volo.Abp.Users;
 using Volo.Abp.VirtualFileSystem;
 using System.IdentityModel.Tokens.Jwt;
+using VCareer.HttpApi.Host.Swagger;
+
 
 namespace VCareer;
 
@@ -77,22 +80,22 @@ public class VCareerHttpApiHostModule : AbpModule
         var hostingEnvironment = context.Services.GetHostingEnvironment();
         var configuration = context.Services.GetConfiguration();
 
-        PreConfigure<OpenIddictBuilder>(builder =>
-        {
-            builder.AddValidation(options =>
-            {
-                options.AddAudiences("VCareer");
-                options.UseLocalServer();
-                options.UseAspNetCore();
-            });
-        });
+           PreConfigure<OpenIddictBuilder>(builder =>
+           {
+               builder.AddValidation(options =>
+               {
+                   options.AddAudiences("VCareer");
+                   options.UseLocalServer();
+                   options.UseAspNetCore();
+               });
+           });
 
-        if (!hostingEnvironment.IsDevelopment())
-        {
-            PreConfigure<AbpOpenIddictAspNetCoreOptions>(options =>
-            {
-                options.AddDevelopmentEncryptionAndSigningCertificate = false;
-            });
+           if (!hostingEnvironment.IsDevelopment())
+           {
+               PreConfigure<AbpOpenIddictAspNetCoreOptions>(options =>
+               {
+                   options.AddDevelopmentEncryptionAndSigningCertificate = false;
+               });
 
             PreConfigure<OpenIddictServerBuilder>(serverBuilder =>
             {
@@ -138,6 +141,7 @@ public class VCareerHttpApiHostModule : AbpModule
         ConfigureJwtOptions(configuration);
         ConfigureBlobStorings(context); //đăng kí cho lưu trữ file blob
         ConfigureGoogleOptions(configuration);
+        ConfigureFilePolicyConfigs(configuration); // Bind FilePolicyConfigs từ appsettings.json
         //  ConfigureClaims();
     }
 
@@ -153,6 +157,13 @@ public class VCareerHttpApiHostModule : AbpModule
     {
         Configure<VCareer.OptionConfigs.GoogleOptions>(configuration.GetSection("Authentication:Google"));
     }
+
+    private void ConfigureFilePolicyConfigs(IConfiguration configuration)
+    {
+        Configure<VCareer.Constants.FilePolicy.FilePolicyConfigs>(configuration.GetSection("FileBlobStorageConfig"));
+    }
+
+
     private void ConfigureAuthentication(ServiceConfigurationContext context, IConfiguration configuration)
     {
         // KHÔNG ForwardIdentityAuthenticationForBearer vì nó sẽ force tất cả JWT Bearer authentication 
@@ -162,9 +173,17 @@ public class VCareerHttpApiHostModule : AbpModule
         context.Services.Configure<AbpClaimsPrincipalFactoryOptions>(options =>
         {
             options.IsDynamicClaimsEnabled = true;
+
+            // ✅ QUAN TRỌNG: Map claims từ JWT sang ABP format
+            options.ClaimsMap["sub"] = new List<string> { AbpClaimTypes.UserId };
+            options.ClaimsMap[ClaimTypes.NameIdentifier] = new List<string> { AbpClaimTypes.UserId };
+            options.ClaimsMap["role"] = new List<string> { AbpClaimTypes.Role };
+            options.ClaimsMap[ClaimTypes.Role] = new List<string> { AbpClaimTypes.Role };
+            options.ClaimsMap["email"] = new List<string> { AbpClaimTypes.Email };
+            options.ClaimsMap[ClaimTypes.Email] = new List<string> { AbpClaimTypes.Email };
+            options.ClaimsMap["preferred_username"] = new List<string> { AbpClaimTypes.UserName };
         });
 
-        //config DI token generator 
         context.Services.AddTransient<ITokenGenerator, JwtTokenGenerator>();
 
         // Cấu hình Authorization để đảm bảo [Authorize] hoạt động đúng
@@ -466,6 +485,17 @@ public class VCareerHttpApiHostModule : AbpModule
                 options.DocInclusionPredicate((docName, description) => true);
                 options.CustomSchemaIds(type => type.FullName);
 
+                // Configure Swagger để handle file uploads (IFormFile)
+                // Map IFormFile to binary string for Swagger
+                options.MapType<Microsoft.AspNetCore.Http.IFormFile>(() => new OpenApiSchema
+                {
+                    Type = "string",
+                    Format = "binary"
+                });
+                
+                // Add OperationFilter để handle IFormFile trong DTOs
+                options.OperationFilter<FileUploadOperationFilter>();
+
                 // Thêm JWT Bearer Authentication vào Swagger
                 options.AddSecurityDefinition("Bearer", new OpenApiSecurityScheme
                 {
@@ -542,6 +572,10 @@ public class VCareerHttpApiHostModule : AbpModule
         }
 
         app.UseRouting();
+
+        // Enable static files serving from wwwroot
+        app.UseStaticFiles();
+
         app.MapAbpStaticAssets();
         app.UseAbpStudioLink();
         app.UseAbpSecurityHeaders();
