@@ -1,12 +1,9 @@
-/*using System;
+using System;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
-using Volo.Abp;
 using Volo.Abp.Application.Dtos;
 using VCareer.Application.Contracts.Applications;
-using VCareer.Application.Contracts.Permission;
 using VCareer.Controllers;
 
 namespace VCareer.HttpApi.Controllers
@@ -16,7 +13,7 @@ namespace VCareer.HttpApi.Controllers
     /// </summary>
     [ApiController]
     [Route("api/applications")]
-    [Authorize(VCareerPermission.Application.Default)]
+    [Authorize]
     public class ApplicationController : VCareerController
     {
         private readonly IApplicationAppService _applicationAppService;
@@ -27,30 +24,27 @@ namespace VCareer.HttpApi.Controllers
         }
 
         /// <summary>
-        /// Nộp đơn ứng tuyển với CV từ thư viện
+        /// Nộp đơn ứng tuyển với CV online (CandidateCv)
         /// </summary>
-        [HttpPost("apply-with-library-cv")]
-        [Authorize(VCareerPermissions.Application.Apply)]
-        public async Task<ApplicationDto> ApplyWithLibraryCVAsync([FromBody] ApplyWithLibraryCVDto input)
+        [HttpPost("apply-with-online-cv")]
+        public async Task<ApplicationDto> ApplyWithOnlineCVAsync([FromBody] ApplyWithOnlineCVDto input)
         {
-            return await _applicationAppService.ApplyWithLibraryCVAsync(input);
+            return await _applicationAppService.ApplyWithOnlineCVAsync(input);
         }
 
         /// <summary>
-        /// Nộp đơn ứng tuyển với CV tải lên mới
+        /// Nộp đơn ứng tuyển với CV đã tải lên (UploadedCv)
         /// </summary>
-        [HttpPost("apply-with-upload-cv")]
-        [Authorize(VCareerPermissions.Application.Apply)]
-        public async Task<ApplicationDto> ApplyWithUploadCVAsync([FromForm] ApplyWithUploadCVDto input)
+        [HttpPost("apply-with-uploaded-cv")]
+        public async Task<ApplicationDto> ApplyWithUploadedCVAsync([FromBody] ApplyWithUploadedCVDto input)
         {
-            return await _applicationAppService.ApplyWithUploadCVAsync(input);
+            return await _applicationAppService.ApplyWithUploadedCVAsync(input);
         }
 
         /// <summary>
         /// Lấy danh sách đơn ứng tuyển
         /// </summary>
         [HttpGet]
-        [Authorize(VCareerPermissions.Application.View)]
         public async Task<PagedResultDto<ApplicationDto>> GetApplicationListAsync([FromQuery] GetApplicationListDto input)
         {
             return await _applicationAppService.GetApplicationListAsync(input);
@@ -60,7 +54,6 @@ namespace VCareer.HttpApi.Controllers
         /// Lấy thông tin chi tiết đơn ứng tuyển
         /// </summary>
         [HttpGet("{id}")]
-        [Authorize(VCareerPermissions.Application.View)]
         public async Task<ApplicationDto> GetApplicationAsync(Guid id)
         {
             return await _applicationAppService.GetApplicationAsync(id);
@@ -70,7 +63,6 @@ namespace VCareer.HttpApi.Controllers
         /// Cập nhật trạng thái đơn ứng tuyển (cho nhà tuyển dụng)
         /// </summary>
         [HttpPut("{id}/status")]
-        [Authorize(VCareerPermissions.Application.Manage)]
         public async Task<ApplicationDto> UpdateApplicationStatusAsync(Guid id, [FromBody] UpdateApplicationStatusDto input)
         {
             return await _applicationAppService.UpdateApplicationStatusAsync(id, input);
@@ -80,7 +72,6 @@ namespace VCareer.HttpApi.Controllers
         /// Hủy đơn ứng tuyển (cho ứng viên)
         /// </summary>
         [HttpPut("{id}/withdraw")]
-        [Authorize(VCareerPermissions.Application.Withdraw)]
         public async Task<ApplicationDto> WithdrawApplicationAsync(Guid id, [FromBody] WithdrawApplicationDto input)
         {
             return await _applicationAppService.WithdrawApplicationAsync(id, input);
@@ -90,7 +81,6 @@ namespace VCareer.HttpApi.Controllers
         /// Đánh dấu đã xem đơn ứng tuyển
         /// </summary>
         [HttpPut("{id}/mark-viewed")]
-        [Authorize(VCareerPermissions.Application.Manage)]
         public async Task<ApplicationDto> MarkAsViewedAsync(Guid id)
         {
             return await _applicationAppService.MarkAsViewedAsync(id);
@@ -100,8 +90,7 @@ namespace VCareer.HttpApi.Controllers
         /// Lấy thống kê đơn ứng tuyển
         /// </summary>
         [HttpGet("statistics")]
-        [Authorize(VCareerPermissions.Application.Statistics)]
-        public async Task<ApplicationStatisticsDto> GetApplicationStatisticsAsync([FromQuery] Guid? jobId = null, [FromQuery] Guid? companyId = null)
+        public async Task<ApplicationStatisticsDto> GetApplicationStatisticsAsync([FromQuery] Guid? jobId = null, [FromQuery] int? companyId = null)
         {
             return await _applicationAppService.GetApplicationStatisticsAsync(jobId, companyId);
         }
@@ -110,7 +99,6 @@ namespace VCareer.HttpApi.Controllers
         /// Lấy danh sách đơn ứng tuyển của ứng viên
         /// </summary>
         [HttpGet("my-applications")]
-        [Authorize(VCareerPermissions.Application.View)]
         public async Task<PagedResultDto<ApplicationDto>> GetMyApplicationsAsync([FromQuery] GetApplicationListDto input)
         {
             return await _applicationAppService.GetMyApplicationsAsync(input);
@@ -120,7 +108,6 @@ namespace VCareer.HttpApi.Controllers
         /// Lấy danh sách đơn ứng tuyển của công ty
         /// </summary>
         [HttpGet("company-applications")]
-        [Authorize(VCareerPermissions.Application.Manage)]
         public async Task<PagedResultDto<ApplicationDto>> GetCompanyApplicationsAsync([FromQuery] GetApplicationListDto input)
         {
             return await _applicationAppService.GetCompanyApplicationsAsync(input);
@@ -130,17 +117,15 @@ namespace VCareer.HttpApi.Controllers
         /// Lấy danh sách đơn ứng tuyển cho một công việc cụ thể
         /// </summary>
         [HttpGet("job/{jobId}")]
-        [Authorize(VCareerPermissions.Application.Manage)]
         public async Task<PagedResultDto<ApplicationDto>> GetJobApplicationsAsync(Guid jobId, [FromQuery] GetApplicationListDto input)
         {
             return await _applicationAppService.GetJobApplicationsAsync(jobId, input);
         }
 
         /// <summary>
-        /// Tải xuống CV của đơn ứng tuyển
+        /// Tải xuống CV của đơn ứng tuyển (PDF hoặc render HTML)
         /// </summary>
         [HttpGet("{id}/download-cv")]
-        [Authorize(VCareerPermissions.Application.DownloadCV)]
         public async Task<IActionResult> DownloadApplicationCVAsync(Guid id)
         {
             var fileBytes = await _applicationAppService.DownloadApplicationCVAsync(id);
@@ -153,12 +138,9 @@ namespace VCareer.HttpApi.Controllers
         /// Xóa đơn ứng tuyển (soft delete)
         /// </summary>
         [HttpDelete("{id}")]
-        [Authorize(VCareerPermissions.Application.Delete)]
         public async Task DeleteApplicationAsync(Guid id)
         {
             await _applicationAppService.DeleteApplicationAsync(id);
         }
     }
 }
-
-*/
